feat:SCRUM-95 온보딩화면 UI조정 및 VM 추가 및 속성이동
